+++ internal/platform/id/id.go
@@ -1,0 +1,7 @@
+package id
+
+import "github.com/google/uuid"
+
+func New() uuid.UUID {
+	return uuid.Must(uuid.NewV7())
+}